# internal/rest/openapi3.yaml
components:
  requestBodies:
    CreateTasksRequest:
      content:
        application/json:
          schema:
            properties:
              task:
                $ref: '#/components/schemas/Task'
      description: Request used for creating a task.
      required: true
  responses:
    CreateTasksResponse:
      content:
        application/json:
          schema:
            properties:
              success:
                type: boolean
      description: Response returned back after creating tasks.
    ErrorResponse:
      content:
        application/json:
          schema:
            properties:
              error:
                type: string
      description: Response when errors happen.
    ReadAllTasksResponse:
      content:
        application/json:
          schema:
            items:
              properties:
                task:
                  $ref: '#/components/schemas/Task'
              type: object
            type: array
      description: Response returned back after searching one task.
    ReadTaskByIdResponse:
      content:
        application/json:
          schema:
            properties:
              task:
                $ref: '#/components/schemas/Task'
      description: Response returned back after searching one task.
  schemas:
    Task:
      properties:
        content:
          type: string
        id:
          format: uuid
          type: string
        timestamp:
          format: int64
          type: integer
        title:
          type: string
        views:
          format: int64
          type: integer
      type: object
info:
  contact:
    url: shay_soifer@dell.com
  description: REST APIs used for interacting with the Edge BE onboarding Service
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Edge BE onboarding exercise - API
  version: 0.0.1
openapi: 3.0.0
paths:
  /tasks:
    get:
      operationId: GetAllTasks
      responses:
        "200":
          $ref: '#/components/responses/ReadAllTasksResponse'
        "400":
          $ref: '#/components/responses/ErrorResponse'
        "500":
          $ref: '#/components/responses/ErrorResponse'
      tags:
      - Task management
    post:
      operationId: CreateTask
      requestBody:
        $ref: '#/components/requestBodies/CreateTasksRequest'
      responses:
        "201":
          $ref: '#/components/responses/CreateTasksResponse'
        "400":
          $ref: '#/components/responses/ErrorResponse'
        "500":
          $ref: '#/components/responses/ErrorResponse'
      tags:
      - Task management
  /tasks/{taskId}:
    get:
      operationId: ReadTask
      parameters:
      - in: path
        name: taskId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          $ref: '#/components/responses/ReadTaskByIdResponse'
        "500":
          $ref: '#/components/responses/ErrorResponse'
      tags:
      - Task management
servers:
- description: Local development
  url: http://0.0.0.0:9234
tags:
- description: Task management services
  name: Task management